+++ src/components/atoms/card/card.module.scss
@@ -6,7 +6,6 @@
   box-shadow: 0 0 4px vars.$drop-shadow-small, 0 4px 16px vars.$drop-shadow-small;
   background-color: vars.$bg-primary;
   padding: 32px;
-  // min-width: fit-content;
 
   &NoPadding {
     padding: 0;
