+++ src/components/atoms/index.js
@@ -1,4 +1,3 @@
-import Icon from './icon/icon';
-import Heading from './heading/heading';
-
-export { Icon, Heading };
+export { default as Card } from './card/card';
+export { default as Icon } from './icon/icon';
+export { default as Heading } from './heading/heading';
