+++ src/components/molecules/index.js
@@ -1,7 +1,7 @@
-import Tag from './tag/tag';
-import Button from './button/button';
-import TextWithIcon from './text-with-icon/text-with-icon';
-import Checkbox from './checkbox/checkbox';
-import HeaderLink from './header-link/header-link';
-
-export { Tag, TextWithIcon, Button, Checkbox, HeaderLink };
+export { default as Accordion } from './accordion/accordion';
+export { default as Button } from './button/button';
+export { default as Checkbox } from './checkbox/checkbox';
+export { default as HeaderLink } from './header-link/header-link';
+export { default as Input } from './input/input';
+export { default as Tag } from './tag/tag';
+export { default as TextWithIcon } from './text-with-icon/text-with-icon';
