+++ src/components/organisms/account-data/account-data.module.scss
@@ -1,5 +1,4 @@
-@use '../../../styles/color-variables.scss';
-@use '../../../styles/_font-variables';
+@use 'src/styles/variables' as vars;
 
 .accountData {
   max-width: 648px;
@@ -7,15 +6,7 @@
   box-sizing: border-box;
   display: flex;
   flex-direction: column;
-  font-family: font-variables.$mainFont;
-}
-
-.title {
-  margin: 0;
-  color: color-variables.$dark-primary;
-  font-weight: 700;
-  font-size: 24px;
-  line-height: calc(28 / 24);
+  font-family: vars.$mainFont;
 }
 
 .form {
