+++ src/components/organisms/account-overview/account-overview.module.scss
@@ -1,3 +1,5 @@
+@use 'src/styles/variables' as vars;
+
 .accountOverview {
   max-width: 312px;
   width: 100%;
@@ -5,6 +7,8 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  font-size: vars.$font-size-m;
+  line-height: vars.$line-height-m;
 }
 
 .avatar {
@@ -22,8 +26,6 @@
 
 .personalData {
   font-weight: 700;
-  font-size: 18px;
-  line-height: calc(24 / 18);
   margin: 0 0 32px;
 }
 
@@ -32,8 +34,6 @@
   padding: 0;
   list-style: none;
   margin: 0 0 32px;
-  font-size: 18px;
-  line-height: calc(24 / 18);
   width: 100%;
 
   &Item {
