+++ src/components/organisms/course-benefits/course-benefits.module.scss
@@ -7,38 +7,41 @@
   padding: 32px;
   font-family: vars.$mainFont;
 }
-.title > button {
-    margin: 0;
-    font-weight: 700;
-    font-size: vars.$font-size-l;
-    line-height: vars.$line-height-l;
 
+.title > button {
+  margin: 0;
+  font-weight: 700;
+  font-size: vars.$font-size-l;
+  line-height: vars.$line-height-l;
 }
+
 .icon {
   max-width: 20px;
   max-height: 20px;
   margin-right: 10px;
 }
+
 .heading {
   display: flex;
   align-items: center;
 }
+
 .benefitsList {
   max-width: 808px;
   display: grid;
   grid-template-columns: repeat(2, 392px);
   gap: 32px 24px;
-  margin: 0;
   padding: 0;
-  margin-top: 32px;
+  margin: 32px 0 0;
 }
+
 .benefit {
   max-width: 392px;
   display: flex;
   flex-direction: column;
   list-style-type: none;
 }
+
 .text {
-  margin: 0;
-  margin-top: 16px;
+  margin: 16px 0 0;
 }
