+++ src/components/organisms/course-completed/course-completed.module.scss
@@ -24,8 +24,7 @@
     .text {
       font-size: vars.$font-size-m;
       line-height: vars.$line-height-m;
-      margin: 0;
-      margin-bottom: 18px;
+      margin: 0 0 18px;
       text-align: center;
 
       .link {
