+++ src/components/organisms/course-description/course-description.module.scss
@@ -1,5 +1,4 @@
-@import 'src/styles/color-variables';
-@import 'src/styles/_font-variables';
+@use 'src/styles/variables' as vars;
 
 .description {
   max-width: 872px;
@@ -11,14 +10,14 @@
 
 .title > button {
   font-weight: 700;
-  font-size: $font-size-l;
-  line-height: $line-height-l;
+  font-size: vars.$font-size-l;
+  line-height: vars.$line-height-l;
 }
 
 .text {
-  font-size: $font-size-m;
-  line-height: $line-height-m;
-  color: $dark-primary;
+  font-size: vars.$font-size-m;
+  line-height: vars.$line-height-m;
+  color: vars.$dark-primary;
   margin: 32px 0 0;
 }
 
@@ -33,7 +32,7 @@
 
 .task {
   margin-top: 18px;
-  color: $dark-primary;
-  font-size: $font-size-m;
-  line-height: $line-height-m;
+  color: vars.$dark-primary;
+  font-size: vars.$font-size-m;
+  line-height: vars.$line-height-m;
 }
