+++ src/components/organisms/course-preview/course-preview.jsx
@@ -1,8 +1,7 @@
 import PropTypes from 'prop-types';
 import { useNavigate } from 'react-router-dom';
+import { Card, Heading } from '../../atoms';
 import { Button, Tag, TextWithIcon } from '../../molecules';
-import Card from '../../templates/card/card';
-import { Heading } from '../../atoms';
 import getDeclensionOf from '../../../utils/getDeclensionOf';
 import styles from './course-preview.module.scss';
