+++ src/components/organisms/test/test.jsx
@@ -1,13 +1,12 @@
 import PropTypes from 'prop-types';
-import { Heading } from '../../atoms';
+import { Card, Heading } from '../../atoms';
 import { Button } from '../../molecules';
-import { Card } from '../../templates';
 import styles from './test.module.scss';
 
 function Test({ toggleRender }) {
   return (
     <Card htmlTag="article" className={styles.test}>
-      <Heading level={1} size="l" title="Тест" />
+      <Heading level={2} size="l" title="Тест" />
       <Button view="text" onClick={toggleRender}>
         Посмотреть условия
       </Button>
