+++ src/components/organisms/video-lesson/video-lesson.jsx
@@ -1,18 +1,15 @@
-import { Heading } from '../../atoms';
-import { Card } from '../../templates';
-
+import { Card, Heading } from '../../atoms';
 import styles from './video-lesson.module.scss';
 
 function VideoLesson() {
   return (
     <Card className={styles.container}>
-      <Heading level={1} size="l" className={styles.heading}>
+      <Heading level={2} size="l" className={styles.heading}>
         Видео
       </Heading>
-      <div className={styles.videoContainer}>
+      <div>
         <iframe
           className={styles.video}
-          height="456"
           src="https://www.youtube.com/embed/elDsY6yS9H8"
           title="YouTube video player"
           allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
