+++ src/components/organisms/video-lesson/video-lesson.module.scss
@@ -9,6 +9,7 @@
 .video {
   border: none;
   width: 100%;
+  height: 456px;
   max-width: 808px;
 }
