+++ src/components/templates/index.js
@@ -1,7 +1,2 @@
-import Accordion from './accordion/accordion';
-import Card from './card/card';
-import Header from './header/header';
-import Layout from './layout/layout';
-import NavigationButtons from './navigation-buttons/navigation-buttons';
-
-export { Accordion, Card, Header, Layout, NavigationButtons };
+export { default as Header } from './header/header';
+export { default as Layout } from './layout/layout';
